System/server/requirements.txt
aiofiles==0.6.0
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyio==3.7.1
argon2-cffi==19.1.0
attrs==20.3.0
autocfg==0.0.8
autogluon.common==0.5.2
autogluon-contrib-nlp==0.0.1b20220208
autogluon.core==0.5.2
autogluon.features==0.5.2
autogluon.multimodal==0.5.2
autogluon.tabular==0.5.2
autogluon.text==0.5.2
autogluon.timeseries==0.5.2
autogluon.vision==0.5.2
bcrypt==3.2.0
blis==0.7.8
boto3==1.24.61
brotlipy==0.7.0
catalogue==2.0.8
catboost==1.0.6
certifi==2023.11.17
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
contextvars==2.4
convertdate==2.4.0
cryptography==3.3.1
cymem==2.0.6
Cython==0.29.32
dask==2021.11.2
Deprecated==1.2.13
distributed==2021.11.2
distro==1.8.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fairscale==0.4.6
fastai==2.7.9
fastapi==0.103.2
fastcore==1.5.22
fastdownload==0.0.7
fastprogress==1.0.3
flake8==3.9.2
flit_core==3.6.0
future==0.18.2
gevent==21.12.0
gluoncv==0.10.5.post0
gluonts==0.9.9
greenlet==1.1.3.post0
grpcio==1.43.0
h11==0.14.0
hijri-converter==2.2.4
holidays==0.15
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.9.1
hyperopt==0.2.7
idna==2.10
imageio==2.21.1
immutables==0.18
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==2.11.2
joblib==1.3.2
korean-lunar-calendar==0.2.1
langcodes==3.3.0
lightgbm==3.3.2
llvmlite==0.39.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==2.0.4
msgpack==1.0.4
murmurhash==1.0.8
nbconvert==7.2.2
networkx==2.6.3
nlpaug==1.1.10
nltk==3.7
nptyping==1.4.4
numba==0.56.0
numpy==1.21.6
omegaconf==2.1.2
openai==1.6.0
opencv-python==4.6.0.66
orjson==3.9.7
packaging==20.8
pandas==1.3.5
passlib==1.7.4
pathy==0.6.2
patsy==0.5.2
Pillow==9.2.0
pip==22.3.1
plotly==5.10.0
pluggy==0.13.1
pmdarima==1.8.5
portalocker==2.5.1
preshed==3.0.7
protobuf==3.18.1
py==1.10.0
py4j==0.10.9.7
pyarrow==9.0.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.15.0
pydantic==2.5.2
pydantic_core==2.14.5
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pyflakes==2.3.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.2.0
python-multipart==0.0.6
pytorch-metric-learning==1.3.2
pytz==2023.4
PyWavelets==1.3.0
PyYAML==6.0.1
ray==1.13.0
regex==2022.8.17
requests==2.25.1
rsa==4.6
sacrebleu==2.2.0
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.95
setuptools==65.6.3
six==1.15.0
sktime==0.11.4
smart-open==5.2.1
sniffio==1.3.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==2.0.23
srsly==2.4.4
starlette==0.27.0
statsmodels==0.13.2
tabulate==0.8.10
tbats==1.1.0
tenacity==8.0.1
tensorboardX==2.5.1
thinc==8.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.5.4
tokenizers==0.12.1
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.7.3
torchtext==0.13.1
torchvision==0.13.1
tqdm==4.66.1
transformers==4.20.1
typer==0.4.2
typing_extensions==4.7.1
typish==1.9.3
ujson==5.7.0
urllib3==1.26.2
uvicorn==0.22.0
wasabi==0.10.1
watchfiles==0.20.0
websocket==0.2.1
websocket-client==0.56.0
websockets==11.0.3
wheel==0.38.4
win-inet-pton==1.1.0
wincertstore==0.2
xgboost==1.4.2
yacs==0.1.8
zipp==3.4.0
zope.event==5.0
zope.interface==5.4.0
